Dependecy issues solved, tests now running properly.
