+++ docs/source/index.rst
@@ -61,25 +61,13 @@ Pyeo is divided into high-level and low-level fucntions.
 Function reference
 ==================
 
-.. module:: pyeo
+.. automodule:: pyeo
+   :members:
 
 At present, all processing code is located in pyeo/core.py.
 A small test suite is located in pyeo/tests/pyeo_tests.py; this is designed for use with py.test.
 Some example applications and demos are in pyeo/apps; for an illustration of the use of the library,
 pyeo/apps/change_detection/simple_s2_change_detection.py is recommended.
-
-
-.. automodule:: pyeo.array_utilities
-.. automodule:: pyeo.classification
-.. automodule:: pyeo.coordinate_manipulation
-.. automodule:: pyeo.exceptions
-.. automodule:: pyeo.filesystem_utilities
-.. automodule:: pyeo.masks
-.. automodule:: plotting
-.. automodule:: queries_and_downloads
-.. automodule:: raster_manipulation
-.. automodule:: sen2_funcs
-.. automodule:: validation
 
 Example scripts
 ===============
@@ -89,7 +77,7 @@ simple_s2_change_detection.py
  .. automodule:: pyeo.apps.change_detection.simple_s2_change_detection
 
 create_model_from_shapefile_and_raster.py
-------------------------------
+-----------------------------------------
  .. automodule:: pyeo.apps.model_creation.create_model_from_shapefile_and_raster
 
 composite_directory.py
