+++ pytest.ini
@@ -1,0 +1,9 @@
+[pytest]
+log_file = pyeo/tests/test_logs.log
+log_file_format = %(asctime)s %(levelname)s %(message)s
+log_file_date_format = %Y-%m-%d %H:%M:%S
+log_level = INFO
+
+python_files = test_*.py
+
+testpaths = pyeo/tests
